fix: break class dependency cycle
Extracted creation methods of VitruviusChangeResolver implementations
into the new singleton factory class VitruviusChangeResolverFactory.
