+++ changederivation/src/main/xtend/tools/vitruv/change/changederivation/AbstractStateBasedChangeResolutionStrategy.xtend
@@ -17,6 +17,7 @@ import org.eclipse.emf.ecore.resource.Resource
 import org.eclipse.emf.ecore.resource.impl.ResourceSetImpl
 import org.eclipse.emf.ecore.util.EcoreUtil
 import tools.vitruv.change.composite.description.VitruviusChangeResolver
+import tools.vitruv.change.composite.description.VitruviusChangeResolverFactory
 import tools.vitruv.change.composite.recording.ChangeRecorder
 
 import static com.google.common.base.Preconditions.checkArgument
@@ -80,7 +81,7 @@ abstract class AbstractStateBasedChangeResolutionStrategy implements StateBasedC
             changeRecorder.addToRecording(resource)
             function.apply()
             val recordedChanges = changeRecorder.endRecording
-            val changeResolver = VitruviusChangeResolver.forHierarchicalIds(resource.resourceSet)
+            val changeResolver = VitruviusChangeResolverFactory.forHierarchicalIds(resource.resourceSet)
             return changeResolver.assignIds(recordedChanges)
         }
     }
