+++ changederivation/src/main/xtend/tools/vitruv/change/changederivation/persistence/DeltaBasedResource.xtend
@@ -13,6 +13,7 @@ import tools.vitruv.change.atomic.EChange
 import tools.vitruv.change.atomic.hid.HierarchicalId
 import tools.vitruv.change.composite.description.VitruviusChangeFactory
 import tools.vitruv.change.composite.description.VitruviusChangeResolver
+import tools.vitruv.change.composite.description.VitruviusChangeResolverFactory
 
 import static edu.kit.ipd.sdq.commons.util.org.eclipse.emf.ecore.resource.ResourceSetUtil.withGlobalFactories
 
@@ -36,7 +37,7 @@ class DeltaBasedResource extends ResourceImpl {
 
 	override doLoad(InputStream inputStream, Map<?, ?> options) throws IOException {
 		val deltas = loadDeltas(this.URI)
-		VitruviusChangeResolver.forHierarchicalIds(resourceSet).resolveAndApply(
+		VitruviusChangeResolverFactory.forHierarchicalIds(resourceSet).resolveAndApply(
 			VitruviusChangeFactory.getInstance().createTransactionalChange(deltas))
 	}
