+++ testutils/integration/src/main/xtend/tools/vitruv/change/testutils/views/ChangePublishingTestView.xtend
@@ -19,6 +19,7 @@ import tools.vitruv.change.atomic.uuid.UuidResolver
 import tools.vitruv.change.composite.description.PropagatedChange
 import tools.vitruv.change.composite.description.TransactionalChange
 import tools.vitruv.change.composite.description.VitruviusChangeResolver
+import tools.vitruv.change.composite.description.VitruviusChangeResolverFactory
 import tools.vitruv.change.composite.propagation.ChangeableModelRepository
 import tools.vitruv.change.composite.recording.ChangeRecorder
 import tools.vitruv.change.propagation.ChangePropagationSpecification
@@ -72,7 +73,7 @@ class ChangePublishingTestView implements NonTransactionalTestView {
 		this.modelRepository = changeableModelRepository
 		this.delegate = new BasicTestView(persistenceDirectory, resourceSet, userInteraction, uriMode)
 		this.changeRecorder = new ChangeRecorder(resourceSet)
-		this.changeResolver = VitruviusChangeResolver.forUuids(uuidResolver)
+		this.changeResolver = VitruviusChangeResolverFactory.forUuids(uuidResolver)
 		this.uuidResolution = uuidResolution
 		changeRecorder.beginRecording()
 	}
